fix(customers): customer id to int64

+++ customers/customers_repository.go
@@ -7,5 +7,5 @@ import (
 
 type CustomerRepository interface {
 	common.Repository
-	FindCustomerById(ctx context.Context, customerId int) (*Customer, error)
+	FindCustomerById(ctx context.Context, customerId int64) (*Customer, error)
 }
